+++ src/onyx/messaging/aeron.clj
@@ -215,10 +215,10 @@
           subscribers (mapv (fn [stream-id]
                               (start-subscriber! bind-addr port stream-id virtual-peers decompress-f receive-idle-strategy))
                             (range subscriber-count))]
-      (.addShutdownHook (Runtime/getRuntime) 
-                        (Thread. (fn [] 
-                                   (.deleteAeronDirectory ^MediaDriver$Context media-driver-context))))
-
+      (when embedded-driver? 
+        (.addShutdownHook (Runtime/getRuntime) 
+                          (Thread. (fn [] 
+                                     (.deleteAeronDirectory ^MediaDriver$Context media-driver-context)))))
       (assoc component
              :bind-addr bind-addr
              :external-addr external-addr