+++ test/onyx/flow_conditions/flow_conditions_gen_test.clj
@@ -5,6 +5,7 @@
             [clojure.test :refer :all]
             [com.gfredericks.test.chuck :refer [times]]
             [com.gfredericks.test.chuck.clojure-test :refer [checking]]
+            [taoensso.timbre :refer [info error warn trace fatal] :as timbre]
             [onyx.static.validation :as v]
             [onyx.flow-conditions.fc-routing :as r]
             [onyx.peer.task-compile :as c]
@@ -89,7 +90,7 @@
 (deftest nil-flow-conditions
   "No flow condition routes to all downstream tasks"
   (let [downstream [:a :b]
-        route (r/route-data nil nil nil nil downstream)]
+        route (r/route-data {} {:flow-conditions nil :egress-ids downstream} nil nil)]
     (is (= downstream (:flow route)))
     (is (nil? (:action route)))))
 
@@ -98,7 +99,7 @@
   (let [e (ex-info "hih" {})
         wrapped-e (ex-info "" {:exception e})]
     (is (thrown? clojure.lang.ExceptionInfo
-                 (r/route-data nil nil wrapped-e nil [:a :b])))))
+                 (r/route-data {} {:flow-conditions nil :egress-ids [:a :b]} nil wrapped-e)))))
 
 (deftest conj-downstream-tasks-together
   (checking
@@ -113,8 +114,10 @@
     other-downstream-tasks (gen-sized-vector)]
    (let [target-tasks (mapcat :flow/to flow-conditions)
          downstream (into other-downstream-tasks target-tasks)
-         event (c/flow-conditions->event-map {} flow-conditions :a)
-         results (r/route-data event nil nil flow-conditions downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions flow-conditions})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (= (into #{} target-tasks) (into #{} (:flow results))))
      (is (nil? (:action results))))))
 
@@ -134,8 +137,10 @@
                      (false-flow-condition-gen)))]
    (let [flow-conditions (into true-fcs false-fcs)
          downstream (mapcat :flow/to true-fcs false-fcs)
-         event (c/flow-conditions->event-map {} flow-conditions :a)
-         results (r/route-data event nil nil flow-conditions downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions flow-conditions})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (= (into #{} (mapcat :flow/to true-fcs)) (into #{} (:flow results))))
      (is (nil? (:action results))))))
 
@@ -168,8 +173,10 @@
                      (maybe-predicate)))]
    (let [flow-conditions (into (into (into false-1 true-1) false-1) mixed-1)
          downstream (mapcat :flow/to flow-conditions)
-         event (c/flow-conditions->event-map {} flow-conditions :a)
-         results (r/route-data event nil nil flow-conditions downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions flow-conditions})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (= (into #{} (:flow/to (first true-1))) (into #{} (:flow results))))
      (is (nil? (:action results))))))
 
@@ -213,8 +220,10 @@
    (let [all [retry-false retry-true retry-mixed ss exceptions mixed]
          flow-conditions (apply concat all)
          downstream (mapcat :flow/to flow-conditions)
-         event (c/flow-conditions->event-map {} flow-conditions :a)
-         results (r/route-data event nil nil flow-conditions downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions flow-conditions})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (not (seq (:flow results))))
      (is (= :retry (:action results))))))
 
@@ -233,8 +242,10 @@
                        (flow-to-tasks-gen)
                        (false-flow-condition-gen))]))]
    (let [downstream (mapcat :flow/to mixed)
-         event (c/flow-conditions->event-map {} mixed :a)
-         results (r/route-data event nil nil mixed downstream)
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions mixed})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)
          matches (filter :pred/val mixed)
          excluded-keys (mapcat :flow/exclude-keys matches)]
      (is (into #{} excluded-keys) (:exclusions results))
@@ -267,8 +278,10 @@
                      (maybe-predicate)))]
    (let [fcs [false-all true-all mixed-all mixed-none]
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         results (r/route-data event nil nil fcs downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (= (into #{} downstream) (into #{} (:flow/to results))))
      (is (nil? (:action results))))))
 
@@ -299,8 +312,10 @@
                      (maybe-predicate)))]
    (let [fcs [false-none true-none mixed-none mixed]
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         results (r/route-data event nil nil fcs downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)]
      (is (not (seq (:flow/to results))))
      (is (nil? (:action results))))))
 
@@ -330,8 +345,10 @@
                      (maybe-predicate)))]
    (let [fcs [false-xform true-xform mixed]
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         results (r/route-data event nil nil fcs downstream)
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         results (r/route-data event (:onyx.core/compiled event) nil nil)
          xform (:flow/post-transformation (first true-xform))]
      (is (= xform (:post-transformation results)))
      (is (nil? (:action results))))))
@@ -351,8 +368,10 @@
    (let [inner-error (ex-info "original error" {})
          message (ex-info "exception" {:exception inner-error})
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         routes (r/route-data event nil message fcs downstream)]
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         routes (r/route-data event (:onyx.core/compiled event) nil message)]
      (is (= message (r/flow-conditions-transform message routes fcs event))))))
 
 (deftest post-transformation-invocation
@@ -370,15 +389,18 @@
    (let [inner-error (ex-info "original error" {})
          message (ex-info "exception" {:exception inner-error})
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         routes (r/route-data event nil message fcs downstream)
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         compiled (:onyx.core/compiled event)
+         routes (r/route-data event (:onyx.core/compiled event) nil message)
          transformed (reduce dissoc post-transformed-segment (:exclusions routes))]
-     (is (= transformed (r/flow-conditions-transform message routes fcs event))))))
+     (is (= transformed (r/flow-conditions-transform message routes event compiled))))))
 
 (deftest post-transformation-exclusions
   (checking
    "Key exclusions are applied during post transformation"
-   (times 50)
+   (times 15)
    [fcs
     (gen/vector (->> :a
                      (flow-condition-gen)
@@ -386,16 +408,19 @@
                      (maybe-predicate)))]
    (let [segment (zipmap (mapcat :flow/exclude-keys fcs) (repeat 0))
          downstream (mapcat :flow/to fcs)
-         event (c/flow-conditions->event-map {} fcs :a)
-         routes (r/route-data event nil segment fcs downstream)
+         event (c/flow-conditions->event-map {:onyx.core/serialized-task {:egress-ids downstream}
+                                              :onyx.core/task :a
+                                              :onyx.core/flow-conditions fcs})
+         compiled (:onyx.core/compiled event)
+         routes (r/route-data event compiled nil segment)
          exclusions (mapcat :flow/exclude-keys (filter :pred/val fcs))
          filtered-segment (reduce dissoc segment exclusions)]
-     (is (= filtered-segment (r/flow-conditions-transform segment routes fcs event))))))
+     (is (= filtered-segment (r/flow-conditions-transform segment routes event compiled))))))
 
 (deftest none-placed-after-all
   (checking
     ":flow/to :none placed after :flow/to :all"
-    (times 50)
+    (times 15)
     [fcs-all (gen/not-empty
                (gen/vector (->> :a
                                 (flow-condition-gen)
@@ -410,7 +435,7 @@
 (deftest none-placed-before-other
   (checking
     ":flow/to :none placed first if there is no :flow/to :all"
-    (times 50)
+    (times 15)
     [fcs-other (gen/not-empty
                (gen/vector (->> :a
                                 (flow-condition-gen)
@@ -426,13 +451,13 @@
 (deftest short-circuit-placed-before-other
   (checking
     ":flow/short-circuit? true should be placed before other conditions"
-    (times 50)
+    (times 15)
     [fcs-other (gen/not-empty
                  (gen/vector (->> :a
                                   (flow-condition-gen))))
      fcs-short-circuit (gen/not-empty
                         (gen/vector (->> :a
                                          (flow-condition-gen)
-                                         (flow-short-circuit-condition-gen ))))]
+                                         (flow-short-circuit-condition-gen))))]
     (let [fcs (into [] (concat fcs-other fcs-short-circuit))]
       (is (thrown? Exception (v/validate-flow-conditions fcs []))))))