+++ test/onyx/generative/apply_fn_gen_test.clj
@@ -27,7 +27,7 @@
     output (segment-map)]
    (let [f (fn [segment] output)
          event {:onyx.core/batch [input]}
-         rets (t/apply-fn f false event)
+         rets (t/apply-fn {:fn f :bulk? false} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))]
      (is (= input (:root (first tree))))
      (is (= output (:message (first (:leaves (first tree)))))))))
@@ -40,7 +40,7 @@
     output (gen/vector (segment-map))]
    (let [f (fn [segment] output)
          event {:onyx.core/batch [input]}
-         rets (t/apply-fn f false event)
+         rets (t/apply-fn {:fn f :bulk? false} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))]
      (is (= input (:root (first tree))))
      (is (= output (map :message (:leaves (first tree))))))))
@@ -62,7 +62,7 @@
    (let [mapping (zipmap (map :message input) output)
          f (fn [segment] (get mapping segment))
          event {:onyx.core/batch input}
-         rets (t/apply-fn f false event)
+         rets (t/apply-fn {:fn f :bulk? false} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))]
      (is (= input (map :root tree)))
      (is (= output (map :message (mapcat :leaves tree)))))))
@@ -75,7 +75,7 @@
     params (gen/vector (gen/resize 10 gen/any))]
    (let [f (fn [& args] {:result (or (butlast args) [])})
          event {:onyx.core/batch input :onyx.core/params params}
-         rets (t/apply-fn f false event)
+         rets (t/apply-fn {:fn f :bulk? false} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))]
      (is (every? (partial = params) (map :result (map :message (mapcat :leaves tree))))))))
 
@@ -86,7 +86,7 @@
    [input (gen/vector (gen-segment))]
    (let [f (fn [segment] (throw (ex-info "exception" {:val 42})))
          event {:onyx.core/batch input}
-         rets (t/apply-fn f false event)
+         rets (t/apply-fn {:fn f :bulk? false} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))
          messages (map :message (:leaves (first tree)))]
      (is (= input (map :root tree)))
@@ -102,7 +102,7 @@
    (let [called? (atom false)
          f (fn [segment] (reset! called? true) output)
          event {:onyx.core/batch input}
-         rets (t/apply-fn f true event)
+         rets (t/apply-fn {:fn f :bulk? true} event)
          tree (:tree (persistent-results! (:onyx.core/results rets)))]
      (is (= input (map :root tree)))
      (is (= (map :message input) (map :message (mapcat :leaves tree))))
@@ -120,6 +120,6 @@
          peer-config {:onyx.peer/fn-params {task-name fn-params}}
          task-map (merge {:onyx/name task-name} {:onyx/params catalog-param-keys}
                          catalog-params)
-         event (c/task-params->event-map {} peer-config task-map)]
+         event (c/task-params->event-map {:onyx.core/peer-opts peer-config :onyx.core/task-map task-map})]
      (is (= (into fn-params catalog-param-vals)
             (:onyx.core/params event))))))